% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aecfeedr.R
\docType{package}
\name{aecfeedr-package}
\alias{aecfeedr}
\alias{aecfeedr-package}
\title{aecfeedr: Retrieve Australian Federal Election Information and Results}
\description{
Access and read data from the Australian Electoral Commission's
    FTP media feeds.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tobiasziegler/aecfeedr}
  \item Report bugs at \url{https://github.com/tobiasziegler/aecfeedr/issues}
}

}
\author{
\strong{Maintainer}: David Mallard \email{correspondence@davidmallard.id.au}

}
\keyword{internal}
