% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{feed_get_messages}
\alias{feed_get_messages}
\title{Download an AEC feed file and extract its contents}
\usage{
feed_get_messages(dir_url, filename, destpath)
}
\arguments{
\item{dir_url}{A string containing the URL of the directory}

\item{filename}{A string containing the name of the zip file to download}

\item{destpath}{A string specifying the base path to download and extract
files in}
}
\value{
A character vector of the filepaths extracted to, invisibly
}
\description{
This function connects to the FTP site, downloads a zip file from a feed
directory, and extracts the XML message(s) and metadata contained within.
}
