% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{feed_get_url}
\alias{feed_get_url}
\title{Get the URL for an AEC election feed directory}
\usage{
feed_get_url(id, granularity, verbosity, archived = FALSE)
}
\arguments{
\item{id}{The unique numeric ID of the electoral event}

\item{granularity}{Standard or Detailed granularity}

\item{verbosity}{Eml, Light, LightProgress, Preload or Verbose}

\item{archived}{FALSE for a live election feed, TRUE for a past election}
}
\value{
A string containing the URL
}
\description{
Specify the parameters of the election feed you would like to access, and
this function will provide the URL for the corresponding FTP site and
directory.
}
