% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{feed_list_files}
\alias{feed_list_files}
\title{List all files in an AEC feed directory}
\usage{
feed_list_files(url)
}
\arguments{
\item{url}{A string containing the URL of the directory}
}
\value{
A character vector containing the filenames within the directory
}
\description{
This helper function connects to the FTP site and returns the list of files
within the directory.
}
