% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_directory_url}
\alias{get_directory_url}
\title{Get the URL for an AEC election feed directory}
\usage{
get_directory_url(id, granularity, verbosity, archived = FALSE)
}
\arguments{
\item{id}{The unique numeric ID of the electoral event}

\item{granularity}{Standard or Detailed granularity}

\item{verbosity}{Eml, Light, LightProgress, Preload or Verbose}

\item{archived}{FALSE for a live election feed, TRUE for a past election}
}
\value{
A string containing the URL
}
\description{
Description goes here
}
