% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{list_files}
\alias{list_files}
\title{List all files in an AEC feed directory}
\usage{
list_files(url)
}
\arguments{
\item{url}{A string containing the URL of the directory}
}
\value{
The contents of the directory
}
\description{
Description goes here
}
