% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preload.R
\name{read_preload_house_fp}
\alias{read_preload_house_fp}
\title{Read the House information from the Detailed Preload Results XML message}
\usage{
read_preload_house_fp(x)
}
\arguments{
\item{x}{A string, connection or raw vector to be processed by \code{read_xml}}
}
\value{
A list of tibbles containing static and historical information
}
\description{
Description goes here
}
