% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{read_results_detailed_preload_house}
\alias{read_results_detailed_preload_house}
\title{Read the House information from the Detailed Preload Results XML message}
\usage{
read_results_detailed_preload_house(x)
}
\arguments{
\item{x}{A string, connection or raw vector to be processed by \code{read_xml}}
}
\value{
A list of tibbles containing static and historical information
}
\description{
Description goes here
}
