% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{read_results_house_fp}
\alias{read_results_house_fp}
\title{Read the first preference House votes from an XML results message}
\usage{
read_results_house_fp(x)
}
\arguments{
\item{x}{A string, connection or raw vector to be processed by \code{read_xml}}
}
\value{
A tibble containing the first preference results by polling place
}
\description{
Description goes here
}
