% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{read_results_house_fp_by_pollingplace}
\alias{read_results_house_fp_by_pollingplace}
\title{Read the first preference House results by polling place from an XML message}
\usage{
read_results_house_fp_by_pollingplace(x)
}
\arguments{
\item{x}{A string, connection or raw vector to be processed by \code{read_xml}}
}
\value{
A tibble containing the first preference results by polling place
}
\description{
Description goes here
}
