% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_zeroes_with_na}
\alias{replace_zeroes_with_na}
\title{Unpack the data from XML elements contained in a list-column}
\usage{
replace_zeroes_with_na(x, ...)
}
\arguments{
\item{x}{A data frame}

\item{...}{Grouping variables that will be applied to identify whether a
complete group of rows have zero total votes reporting.}
}
\value{
A data frame with zeroes replaced by \code{NA}s
}
\description{
Some elements within AEC feeds contain a value of 0 when results haven't
been received yet. This function checks whether each full group of vote
count elements (e.g., per polling place) has zero values and, if so, changes
them all to NA. Provide grouping variables via the dots (...) in the
function call.
}
