% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{unpack_xml}
\alias{unpack_xml}
\title{Unpack the data from XML elements contained in a list-column}
\usage{
unpack_xml(data, xml_col, formula)
}
\arguments{
\item{data}{A data frame}

\item{xml_col}{The bare (unquoted) name of the list-column
containing XML data}

\item{formula}{A formula specifying how to extract data from each XML node}
}
\value{
A data frame with the unpacked data
}
\description{
Unpack the data from XML elements contained in a list-column
}
