% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{write_db_tables}
\alias{write_db_tables}
\title{Write data processed from a feed into a database}
\usage{
write_db_tables(conn, data, ...)
}
\arguments{
\item{conn}{A \code{DBIConnection`` object, as returned by }dbConnect()`}

\item{data}{A list of tibbles containing data from an AEC feed}

\item{...}{Other parameters passed on to the \code{dbWriteTable} method}
}
\value{
Invisibly returns \code{data}
}
\description{
Specify the parameters of the election feed you would like to access, and
this function will provide the URL for the corresponding FTP site and
directory.
}
