% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{xml_find_attr}
\alias{xml_find_attr}
\title{Retrieve an attribute from the first node that matches an xpath expression}
\usage{
xml_find_attr(x, xpath, attr, ns = xml_ns(x))
}
\arguments{
\item{x}{A document, node or node set}

\item{xpath}{A string containing a xpath (1.0) expression}

\item{attr}{Name of attribute to extract}

\item{ns}{Optionally, a named vector giving prefix-url pairs, as produced
by \code{\link[=xml_ns]{xml_ns()}}. If provided, all names will be explicitly
qualified with the ns prefix, i.e. if the element \code{bar} is defined
in namespace \code{foo}, it will be called \code{foo:bar}. (And
similarly for attributes). Default namespaces must be given an explicit
name. The ns is ignored when using \code{\link[=xml_name<-]{xml_name<-()}} and
\code{\link[=xml_set_name]{xml_set_name()}}.}
}
\value{
A character vector. \code{NA} is used to represent of attributes that
aren't defined.
}
\description{
This is a helper function that uses \code{xml_find_first} to retrieve a node and
then retrieves the specified attribute on that node.
}
\examples{
x <- xml2::read_xml("<a id='foo' />")
xml_find_attr(x, "/a", "id")
}
