% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{xml_find_text}
\alias{xml_find_text}
\title{Extract the text from the first node that matches an xpath expression}
\usage{
xml_find_text(x, xpath, trim = FALSE, ns = xml_ns(x))
}
\arguments{
\item{x}{A document, node or node set}

\item{xpath}{A string containing a xpath (1.0) expression}

\item{trim}{If \code{TRUE} will trim leading and trailiing spaces}

\item{ns}{Optionally, a named vector giving prefix-url pairs, as produced
by \code{\link[=xml_ns]{xml_ns()}}. If provided, all names will be explicitly
qualified with the ns prefix, i.e. if the element \code{bar} is defined
in namespace \code{foo}, it will be called \code{foo:bar}. (And
similarly for attributes). Default namespaces must be given an explicit
name. The ns is ignored when using \code{\link[=xml_name<-]{xml_name<-()}} and
\code{\link[=xml_set_name]{xml_set_name()}}.}
}
\value{
A character vector.
}
\description{
This is a helper function that uses \code{xml_find_first} to retrieve a node and
then extracts the text from that node.
}
\examples{
x <- xml2::read_xml("<a>foo</a>")
xml_find_text(x, "/a")
}
