% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preload.R
\name{read_preload_house_fp}
\alias{read_preload_house_fp}
\title{Read the House information from a Detailed Preload Results XML message}
\usage{
read_preload_house_fp(x)
}
\arguments{
\item{x}{A string, connection or raw vector to be processed by \code{read_xml}}
}
\value{
A list of tibbles containing static and historical information
}
\description{
The AEC's preload data feed provides the details about election contests,
candidates, and polling places, along with the historical (previous
election) results. This function processes the detailed preload message and
produces a set of tibbles that can be used to produce pre-election content
and can be combined with the light version of live results, which only
identifies contests, candidates and polling places by their ID numbers.
}
